Started moving code around
